Feat: implementando tratamento de dados na rota removePlayers

+++ src/modules/players/useCases/removePlayers/removePlayersUseCase.ts
@@ -6,6 +6,12 @@ class RemovePlayersUseCase {
     constructor(private playersRepository: PlayersRepository) {}
 
     async execute(id: string): Promise<void> {
+       const userAlredyExist = await this.playersRepository.findById(id);
+
+       if(!userAlredyExist) {
+        throw new Error("Usuário não encontrado");
+       }
+
        this.playersRepository.removePlayerById(id);
     }
